requirements.txt
asgiref==3.5.2
backports.zoneinfo==0.2.1
boto3==1.26.59
botocore==1.29.59
cachetools==5.2.0
certifi==2022.9.24
charset-normalizer==2.1.1
currencies==2020.12.12
Django==4.1.3
django-environ==0.9.0
django-storages==1.13.1
djangorestframework==3.14.0
fontawesomefree==6.3.0
googleapis-common-protos==1.57.0
grpcio==1.51.1
grpcio-status==1.51.1
idna==3.4
jmespath==1.0.1
proto-plus==1.22.1
protobuf==4.21.10
psycopg2-binary==2.9.5
pyasn1==0.4.8
pyasn1-modules==0.2.8
PyJWT==2.6.0
python-dateutil==2.8.2
python-magic==0.4.27
pytz==2022.7.1
requests==2.28.1
rsa==4.9
s3transfer==0.6.0
six==1.16.0
sqlparse==0.4.3
stripe==5.2.0
urllib3==1.26.13
uWSGI==2.0.21
